Extract FilterGroupItem
